change float finish

+++ src/remote_control/scripts/singleAxisArcadeDrive.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 import rospy
-from std_msgs.msg import Int32
+from std_msgs.msg import Float32
 import time
 import math
 rospy.init_node('singleAxisArcadeDrive', anonymous=True)
@@ -54,10 +54,10 @@ def Ycallback(data):
 
 
 # left and right subscribers
-rospy.Subscriber(XAxisTopic,Int32,Xcallback)
-rospy.Subscriber(YAxisTopic,Int32,Ycallback)
+rospy.Subscriber(XAxisTopic,Float32,Xcallback)
+rospy.Subscriber(YAxisTopic,Float32,Ycallback)
 # left and right publishers
-pub=[rospy.Publisher(leftMotorChannel, Int32, queue_size=10),rospy.Publisher(rightMotorChannel, Int32, queue_size=10)]
+pub=[rospy.Publisher(leftMotorChannel, Float32, queue_size=10),rospy.Publisher(rightMotorChannel, Int32, queue_size=10)]
 print("Mapping {0},{1} ==> {2},{3}".format(XAxisTopic,YAxisTopic,leftMotorChannel,rightMotorChannel))
 while not rospy.is_shutdown():
     #convert into radial coords
